+++ app/src/main/java/org/simple/clinic/registration/phone/RegistrationPhoneScreenKey.kt
@@ -1,6 +1,6 @@
 package org.simple.clinic.registration.phone
 
-import kotlinx.android.parcel.IgnoredOnParcel
+import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.navigation.v2.ScreenKey