+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Fix instant search error views overlapping each other
 - Remove usage of deprecated constructor from `MobiusDelegate`
 - Migrate `OverdueAppointment` to add patients without phone number
+- Migrate `kotlinx.android.parcel` usage to `kotlinx.parcelize`
 
 ### Changes
 - Updated translations: `hi-IN`, `bn-BD`, `te-IN`, `bn-IN`, `am-ET`, `pa-IN`, `ti-ET`, `kn-IN`, `mr-IN`, `ta-IN`, `so-ET`, `om-ET`